+++ src/main/java/com/ftc/demo/entities/Status.java
@@ -1,13 +1,10 @@
 package com.ftc.demo.entities;
 
-import java.util.ArrayList;
-import java.util.List;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
@@ -20,7 +17,7 @@ import lombok.RequiredArgsConstructor;
 public class Status {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
 	@NonNull
 	private String name;
 }
